Refactor mds-api-server module for maintainability
