+++ packages/mds-api-server/middleware/compression.ts
@@ -1,0 +1,6 @@
+import type { CompressionOptions } from 'compression'
+import compression from 'compression'
+
+export type CompressionMiddlewareOptions = Partial<CompressionOptions>
+
+export const CompressionMiddleware = (options: CompressionMiddlewareOptions = {}) => compression(options)
