+++ packages/mds-api-server/middleware/raw-body-parser.ts
@@ -1,0 +1,7 @@
+import type { Options as OptionsRaw } from 'body-parser'
+import bodyParser from 'body-parser'
+
+export type RawBodyParserMiddlewareOptions = Partial<OptionsRaw>
+
+export const RawBodyParserMiddleware = (options: RawBodyParserMiddlewareOptions = {}) =>
+  bodyParser.raw({ limit: '5mb', ...options })
